requirements.txt
groq
streamlitlangdetect